# Speed up reference usage and bump version