--- a/docs/source/conf.py
+++ b/docs/source/conf.py
@@ -9,7 +9,7 @@
 project = "fake-vcf"
 copyright = "2023, Magnus Wahlberg"
 author = "Magnus Wahlberg"
-version = "0.2.0"
+version = "0.2.1"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration