--- a/pyproject.toml
+++ b/pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "fake-vcf"
-version = "0.2.0"
+version = "0.2.1"
 description = "A fake vcf file generator "
 readme = "README.md"
 authors = ["fake-vcf <endast@gmail.com>"]