--- a/tests/test_data/chr1_small.vcf
+++ b/tests/test_data/chr1_small.vcf
@@ -1,5 +1,5 @@
 ##fileformat=VCFv4.2
-##source=VCFake 0.2.0
+##source=VCFake 0.1.0
 ##FILTER=<ID=PASS,Description="All filters passed">
 ##INFO=<ID=NS,Number=1,Type=Integer,Description="Number of Samples With Data">
 ##contig=<ID=chr1>