--- a/tests/test_data/reference/parquet/sequence_metadata.json
+++ b/tests/test_data/reference/parquet/sequence_metadata.json
@@ -1,6 +1,6 @@
 {
   "source_reference_file": "test-reference.fa",
-  "fake-vcf-version": "0.2.0",
+  "fake-vcf-version": "0.2.1",
   "reference_files": {
     "chr1": "fasta_chr1.parquet",
     "chr2": "fasta_chr2.parquet",