## Description

Speeds up generation when using a reference fasta file.

### fake-vcf version: 0.2.0

```
fake-vcf generate -r 100 -s 1000 -o /tmp/dump.vcf
...
299 ms ± 6.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```
fake-vcf generate -r 100 -s 1000 -o /tmp/dump.vcf --reference-dir-path reference_dir
...
13.4 s ± 56.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
### fake-vcf version: 0.2.1
```
fake-vcf generate -r 100 -s 1000 -o /tmp/dump.vcf
...
306 ms ± 5.62 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```
fake-vcf generate -r 100 -s 1000 -o /tmp/dump.vcf
...
1.41 s ± 9.23 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


## Related Issue

[Original issue](https://github.com/endast/fake-vcf/issues/232)


## Type of Change

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

- [X] 🔧 Bug fix (non-breaking change which fixes an issue)
- [X] 🥂 Improvement (non-breaking change which improves an existing feature)